Shared async loading of babel and codemirror
Nice performance boost. Doesn't matter which page you load first that
makes use of the ReactPlayground, it's only loaded once. No more giant
chunks. The difference from doing this and vendoring it into its own
bundle is that if we vendor it then we load it through index.html and
it is loaded on every page. This solution loads it only when needed.